I've been neglecting committing, but a lot of progress on the qt transition here. using qt thread for exp thread, but still seems like some minor issues with thread crashing with variables?

+++ Chimera/Source/Agilent/AgilentCore.h
@@ -55,7 +55,7 @@ class AgilentCore : public IDeviceCore
 		void errorFinish () {};
 	private:
 		deviceOutputInfo expRunSettings;
-		const double sampleRate;
+		const ULONG sampleRate;
 		const std::string memoryLoc;
 		const agilentSettings initSettings;
 		// not that important, just used to check that number of triggers in script matches number in agilent.
